Package reorganization (#42)
* Iteration 0 pre release (#7)

* PR Template and .gitignore (#3)

* PR Template and .gitignore

* Add code reviews check list to the PR template

* Team 1 Into - Lab 1 (#2)

* Add README.md file

* tim flucker adding his introduction

* added some formatting for the README.md

* Some message of Xiaobing Hou

* some message of Xiaobing Hou

* some message about xiaobing

* Xiaobing_lab1

* Added intro - Jean Dorancy

* Update README.md

Add Intro

* Minor edits


* Adding swagger doc (#4)

* Add README.md file

* tim flucker adding his introduction

* added some formatting for the README.md

* adding swagger document as project artifact

* updating basepath, and changing paths to use hyphens


* WIP:Initial project (#6)

* Add README.md file

* tim flucker adding his introduction

* added some formatting for the README.md

* adding swagger document as project artifact

* updating basepath, and changing paths to use hyphens

* adding initial project files


* Iteration 0 (#5)

* Iteration 0 Documents

* Minor edit to README

* Added risk management sheet

* Added SPPP document

* Added team presentation for iteration 0

* Add progress report and update risk management

* Removed requirements template from SPPP


* updating pom to include new dependency, removing code with old packages

* re organizing classes to be packaged by feature

* updating unit tests to import from new packages

* re-organizing the test classes to match the new package structure
