--- src/main/java/edu/bu/metcs673/trackr/repo/TrackrUserRepository.java
@@ -1,34 +1,0 @@
-package edu.bu.metcs673.trackr.repo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import edu.bu.metcs673.trackr.domain.TrackrUser;
-
-/**
- * Directly interfaces with the "USERS" table in the H2 repository, using the
- * JPA library.
- * 
- * @author Tim Flucker
- *
- */
-@Repository
-public interface TrackrUserRepository extends JpaRepository<TrackrUser, Long> {
-
-	/**
-	 * Custom JPA query that checks to see if a record with the provided 'username'
-	 * already exists in the DB.
-	 * 
-	 * @param username
-	 * @return boolean
-	 */
-	public boolean existsByUsername(String username);
-
-	/**
-	 * Custom JPA query that finds a user by its 'username' value.
-	 * 
-	 * @param username
-	 * @return
-	 */
-	public TrackrUser findByUsername(String username);
-}
